+++ trans/src/wich/codegen/model/Expr.java
@@ -23,9 +23,6 @@ SOFTWARE.
 */
 package wich.codegen.model;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public abstract class Expr extends OutputModelObject {
-	public List<TmpVarDef> tmpVarDefs = new ArrayList<>();
+//	public List<TmpVarDef> tmpVarDefs = new ArrayList<>();
 }
