+++ trans/src/wich/codegen/model/RefCountREF.java
@@ -1,0 +1,9 @@
+package wich.codegen.model;
+
+public class RefCountREF extends Stat {
+	public String varName;
+
+	public RefCountREF(String varName) {
+		this.varName = varName;
+	}
+}
