+++ trans/src/wich/codegen/model/VarDefStat.java
@@ -23,9 +23,6 @@ SOFTWARE.
 */
 package wich.codegen.model;
 
-import java.util.ArrayList;
-import java.util.List;
-
 /** Represents a var x = <expr> statement. For globals, the C templates split
  *  these model objects into def and init so generated functions can see them.
  */
@@ -33,8 +30,6 @@ public class VarDefStat extends Stat {
 	public final String name;
 	@ModelElement public WichType type;
 	@ModelElement public Expr expr;
-	@ModelElement public List<TmpVarDef> localTemps = new ArrayList<>();
-	public String ref = null;
 
 	public VarDefStat(String name) {
 		this.name = name;
